Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 
 int    _atoi(const char *s)
 {
-	long int	i;
+	unsigned int	i;
 	int		result;
 	int		sign;
 	int		j;
